requirements.txt
streamlit 
playwright
pillow
fpdf